# perf: add gzip writer reuse