… godeltaprof/compat/go.sum
… godeltaprof/internal/pprof/gzip_go16.go
… godeltaprof/internal/pprof/gzip_go17.go