https://github.com/grafana/pyroscope-go/issues/162

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/pyroscope-go/godeltaprof
cpu: Apple M2
       │ /Users/grafana/pyroscope-go/bench_before.txt │ /Users/grafana/pyroscope-go/bench_after.txt │
       │                    sec/op                    │       sec/op         vs base                │
Heap-8                                  141.30µ ± 15%          91.69µ ± 16%  -35.11% (p=0.000 n=10)

       │ /Users/grafana/pyroscope-go/bench_before.txt │ /Users/grafana/pyroscope-go/bench_after.txt │
       │                     B/op                     │        B/op          vs base                │
Heap-8                                   953.8Ki ± 3%         130.4Ki ± 15%  -86.33% (p=0.000 n=10)

       │ /Users/grafana/pyroscope-go/bench_before.txt │ /Users/grafana/pyroscope-go/bench_after.txt │
       │                  allocs/op                   │      allocs/op       vs base                │
Heap-8                                     268.5 ± 1%           217.5 ± 16%  -18.99% (p=0.001 n=10)
```